Break Case Bioinformatics Tab into a new file
