+++ src/encoded/static/components/browse/columnExtensionMap.js
@@ -10,11 +10,12 @@ import { basicColumnExtensionMap,
     DisplayTitleColumnWrapper,
     DisplayTitleColumnDefault,
     DisplayTitleColumnUser } from '@hms-dbmi-bgm/shared-portal-components/es/components/browse/components/table-commons';
-import { Schemas, typedefs } from './../util';
 
+import { Schemas, typedefs } from './../util';
+import { generateRelationshipMapping, sortAndAddRolePropsToQCMs } from '../util/item';
 import { variantSampleColumnExtensionMap, structuralVariantSampleColumnExtensionMap, VariantSampleDisplayTitleColumn } from './variantSampleColumnExtensionMap';
 import QuickPopover from '../item-pages/components/QuickPopover';
-import { generateRelationshipMapping, QCMFlag, sortAndAddRolePropsToQCMs } from '../item-pages/CaseView';
+import { QCMFlag } from '../item-pages/components/QCM';
 import { CurrentFamilyController, findCanonicalFamilyIndex } from '../item-pages/CaseView/CurrentFamilyController';
 
 // eslint-disable-next-line no-unused-vars
